refactor(types): seperate types file into specific modules
