+++ src/types/ui.ts
@@ -1,0 +1,7 @@
+interface Props {
+	children: React.ReactNode;
+}
+
+type Theme = "light" | "dark";
+
+export type { Props, Theme };
